Add test appointment management and refactor models
- Introduced `ITestAppointmentRepository` and `TestAppointment` model.
- Updated `Applications` class properties for clarity.
- Changed `FineFees` type in `DetainedLicense` from `decimal` to `float`.
- Removed `Form1` and associated files, reflecting UI changes.
- Added transaction management methods in `DBHelper`.
- Enhanced `ApplicationsRepository` with new methods for local driving license applications.
- Updated `ILicenseDetainRepository` to include a method for detaining licenses.
- Created `TestAppointmentRepository` for handling test appointment data.
- Removed unused methods from `InternationalLicenseController` and `TestTypeController`.
- Updated project files to include new models and repositories.
- Adjusted UI components in `TestManagementForm` for displaying test statistics.

+++ Core/Models/DetainedLicense.cs
@@ -9,7 +9,7 @@ namespace Core.Models
         public int DetainID { get; set; }
         public int LicenseID { get; set; }
         public DateTime DetainDate { get; set; }
-        public decimal FineFees { get; set; }
+        public float FineFees { get; set; }
         public int CreatedByUserID { get; set; }
         public bool IsReleased { get; set; }
         public DateTime? ReleaseDate { get; set; }
